+++ src/Storage/Normalizer/SerializableCompressNormalizer.php
@@ -3,6 +3,8 @@ declare(strict_types=1);
 
 namespace Heptacom\HeptaConnect\Core\Storage\Normalizer;
 
+use Heptacom\HeptaConnect\Core\Storage\Exception\GzipCompressException;
+
 class SerializableCompressNormalizer extends SerializableNormalizer
 {
     public function getType(): string
@@ -12,6 +14,12 @@ class SerializableCompressNormalizer extends SerializableNormalizer
 
     public function normalize($object, ?string $format = null, array $context = [])
     {
-        return \gzcompress(parent::normalize($object, $format, $context));
+        $result = \gzcompress(parent::normalize($object, $format, $context));
+
+        if (!\is_string($result)) {
+            throw new GzipCompressException(1637432095);
+        }
+
+        return $result;
     }
 }